+++ toolium/visual_test.py
@@ -29,13 +29,7 @@ from selenium.common.exceptions import NoSuchElementException
 from toolium.driver_wrappers_pool import DriverWrappersPool
 from toolium.utils.path_utils import get_valid_filename, makedirs_safe
 
-try:
-    from needle.engines.perceptualdiff_engine import Engine as PerceptualEngine
-    from needle.engines.pil_engine import Engine as PilEngine
-    from PIL import Image
-    from needle.engines.imagemagick_engine import Engine as MagickEngine
-except ImportError:
-    pass
+from PIL import Image, ImageChops
 
 
 class VisualTest(object):
@@ -56,8 +50,6 @@ class VisualTest(object):
         self.force = force
         if not self.driver_wrapper.config.getboolean_optional('VisualTests', 'enabled') and not self.force:
             return
-        if 'PerceptualEngine' not in globals():
-            raise Exception('The visual tests are enabled, but needle is not installed')
 
         self.utils = self.driver_wrapper.utils
         self.logger = logging.getLogger(__name__)
@@ -72,7 +64,6 @@ class VisualTest(object):
                                                                          get_valid_filename(baseline_name))
 
         self.baseline_directory = self.driver_wrapper.visual_baseline_directory
-        self.engine = self._get_engine()
         self.save_baseline = self.driver_wrapper.config.getboolean_optional('VisualTests', 'save')
 
         # Create folders
@@ -92,28 +83,6 @@ class VisualTest(object):
             shutil.copyfile(orig_javascript_path, dst_javascript_path)
             shutil.copyfile(orig_css_path, dst_css_path)
             self._add_summary_to_report()
-
-    def _get_engine(self):
-        """Get configured or default visual engine
-
-        :returns: engine instance
-        """
-        engine_type = self.driver_wrapper.config.get_optional('VisualTests', 'engine', 'pil')
-        if engine_type == 'perceptualdiff':
-            engine = PerceptualEngine()
-        elif engine_type == 'imagemagick' and 'MagickEngine' not in globals():
-            self.logger.warning("Engine '%s' not found, using pil instead. You need needle 0.4+ to use this engine.",
-                                engine_type)
-            engine = PilEngine()
-        elif engine_type == 'imagemagick':
-            engine = MagickEngine()
-        elif engine_type == 'pil':
-            engine = PilEngine()
-        else:
-            self.logger.warning("Engine '%s' not found, using pil instead. Review your properties.cfg file.",
-                                engine_type)
-            engine = PilEngine()
-        return engine
 
     def assert_screenshot(self, element, filename, file_suffix=None, threshold=0, exclude_elements=[]):
         """Assert that a screenshot of an element is the same as a screenshot on disk, within a given threshold
@@ -140,11 +109,11 @@ class VisualTest(object):
             except NoSuchElementException as e:
                 self.logger.warning("Element to be excluded not found: %s", str(e))
 
-        baseline_file = os.path.join(self.baseline_directory, '{}.png'.format(filename))
+        baseline_path = os.path.join(self.baseline_directory, '{}.png'.format(filename))
         filename_with_suffix = '{0}__{1}'.format(filename, file_suffix) if file_suffix else filename
         unique_name = '{0:0=2d}_{1}'.format(DriverWrappersPool.visual_number, filename_with_suffix)
         unique_name = '{}.png'.format(get_valid_filename(unique_name))
-        output_file = os.path.join(self.output_directory, unique_name)
+        output_path = os.path.join(self.output_directory, unique_name)
         report_name = '{}<br>({})'.format(file_suffix, filename) if file_suffix else '-<br>({})'.format(filename)
 
         # Get screenshot and modify it
@@ -154,28 +123,29 @@ class VisualTest(object):
         img = self.desktop_resize(img)
         img = self.exclude_elements(img, exclude_web_elements)
         img = self.crop_element(img, web_element)
-        img.save(output_file)
+        img.save(output_path)
         DriverWrappersPool.visual_number += 1
 
         # Determine whether we should save the baseline image
         if self.save_baseline:
             # Copy screenshot to baseline
-            shutil.copyfile(output_file, baseline_file)
+            shutil.copyfile(output_path, baseline_path)
 
             if self.driver_wrapper.config.getboolean_optional('VisualTests', 'complete_report'):
-                self._add_result_to_report('baseline', report_name, output_file, None, 'Screenshot added to baseline')
+                self._add_result_to_report('baseline', report_name, output_path, None, None,
+                                           'Screenshot added to baseline')
 
             self.logger.debug("Visual screenshot '%s' saved in visualtests/baseline folder", filename)
-        elif not os.path.exists(baseline_file):
+        elif not os.path.exists(baseline_path):
             # Baseline should exist if save mode is not enabled
-            error_message = "Baseline file not found: %s" % baseline_file
+            error_message = f'Baseline file not found: {baseline_path}'
             self.logger.warning(error_message)
-            self._add_result_to_report('diff', report_name, output_file, None, 'Baseline file not found')
+            self._add_result_to_report('diff', report_name, output_path, None, None, 'Baseline file not found')
             if self.driver_wrapper.config.getboolean_optional('VisualTests', 'fail') or self.force:
                 raise AssertionError(error_message)
         else:
             # Compare the screenshots
-            self.compare_files(report_name, output_file, baseline_file, threshold)
+            self.compare_files(report_name, output_path, baseline_path, threshold)
 
     def get_scrolls_size(self):
         """Return Chrome and Explorer scrolls sizes if they are visible
@@ -237,7 +207,7 @@ class VisualTest(object):
         scale = img.size[0] / self.utils.get_window_size()['width']
         if scale != 1:
             new_image_size = (int(img.size[0] / scale), int(img.size[1] / scale))
-            img = img.resize(new_image_size, Image.ANTIALIAS)
+            img = img.resize(new_image_size, Image.LANCZOS)
         return img
 
     def get_element_box(self, web_element):
@@ -299,52 +269,110 @@ class VisualTest(object):
 
         return img
 
-    def compare_files(self, report_name, image_file, baseline_file, threshold):
-        """Compare two image files and add result to the html report
+    def compare_files(self, report_name, image_path, baseline_path, threshold):
+        """Compare two image files, generate a new image file with highlighted differences,
+           calculate the percentage of pixels that are different between both images and add result to the html report
 
         :param report_name: name to show in html report
-        :param image_file: image file path
-        :param baseline_file: baseline image file path
+        :param image_path: image file path
+        :param baseline_path: baseline image file path
         :param threshold: percentage threshold
-        :returns: error message
+        :returns: result message
         """
-        width, height = Image.open(image_file).size
-        if isinstance(self.engine, PilEngine):
-            # Pil needs a pixel number threshold instead of a percentage threshold
-            threshold = int(width * height * threshold)
-        try:
-            if 'MagickEngine' in globals() and isinstance(self.engine, MagickEngine):
-                # Workaround: ImageMagick hangs when images are not equal
-                assert (width, height) == Image.open(baseline_file).size, 'Image dimensions do not match'
-            self.engine.assertSameFiles(image_file, baseline_file, threshold)
-            if self.driver_wrapper.config.getboolean_optional('VisualTests', 'complete_report'):
-                self._add_result_to_report('equal', report_name, image_file, baseline_file)
-            return None
-        except AssertionError as exc:
-            diff_message = self._get_diff_message(str(exc), width * height)
-            self._add_result_to_report('diff', report_name, image_file, baseline_file, diff_message)
-            self.logger.warning("Visual error in '%s': %s", os.path.splitext(os.path.basename(baseline_file))[0],
-                                diff_message)
-            if self.driver_wrapper.config.getboolean_optional('VisualTests', 'fail') or self.force:
-                raise exc
-            else:
-                return diff_message
+        # Make two new images with same size
+        with Image.open(image_path) as image:
+            image_size = image.size
+            with Image.open(baseline_path) as baseline:
+                baseline_size = baseline.size
+                max_size = (max(image.width, baseline.width), max(image.height, baseline.height))
+                image_max = Image.new('RGB', max_size)
+                image_max.paste(image.convert('RGB'))
+                baseline_max = Image.new('RGB', max_size)
+                baseline_max.paste(baseline.convert('RGB'))
 
-    def _add_result_to_report(self, result, report_name, image_file, baseline_file, message=''):
+        # Generate and save diff image
+        diff_path = image_path.replace('.png', '.diff.png')
+        diff_pixels_percentage = self.save_differences_image(image_max, baseline_max, diff_path)
+
+        # Check differences and add to report
+        if image_size != baseline_size:
+            # Different size
+            diff_message = f"Image dimensions {image_size} do not match baseline size {baseline_size}"
+            exception_message = f"The new screenshot '{image_path}' size '{image_size}' did not match the" \
+                                f" baseline '{baseline_path}' size '{baseline_size}'"
+            result = 'diff'
+        elif diff_pixels_percentage == 0:
+            # Equal images
+            diff_path = diff_message = None
+            result = 'equal'
+        elif 0 < diff_pixels_percentage <= threshold:
+            # Similar images
+            diff_message = f'Distance is {diff_pixels_percentage:.8f}, less than {threshold} threshold'
+            result = 'equal'
+        else:
+            # Different images
+            diff_message = f'Distance is {diff_pixels_percentage:.8f}, more than {threshold} threshold'
+            exception_message = f"The new screenshot '{image_path}' did not match the baseline '{baseline_path}'" \
+                                f" (by a distance of {diff_pixels_percentage:.8f}, more than {threshold} threshold)"
+            result = 'diff'
+
+        if (result == 'equal' and self.driver_wrapper.config.getboolean_optional('VisualTests', 'complete_report')
+                or result == 'diff'):
+            self._add_result_to_report(result, report_name, image_path, baseline_path, diff_path, diff_message)
+            # Add message to result to be used in unittests
+            result = f'{result}-{diff_message}' if diff_message is not None else result
+
+        if result.startswith('diff'):
+            self.logger.warning(f"Visual error in '{os.path.splitext(os.path.basename(baseline_path))[0]}':"
+                                f" {diff_message}")
+            if self.driver_wrapper.config.getboolean_optional('VisualTests', 'fail') or self.force:
+                raise AssertionError(exception_message)
+
+        return result
+
+    @staticmethod
+    def save_differences_image(image, baseline, diff_path):
+        """Create and save an image showing differences between both images
+
+        :param image: image object
+        :param baseline: reference baseline image object
+        :param diff_path: file path where difference image will be saved
+        :returns: percentage of pixels that are different between both images
+        """
+        # Create a mask with differences
+        mask = ImageChops.difference(image, baseline).convert('L').point(lambda x: 255 if x else 0)
+        # Create a White base
+        white_image = Image.new('RGB', baseline.size, (255, 255, 255))
+        # Add baseline with 50% opacity
+        baseline.putalpha(127)
+        white_image.paste(baseline, (0, 0), baseline)
+        # Add red points in different pixels
+        red_image = Image.new('RGB', baseline.size, (255, 0, 0))
+        white_image.paste(red_image, (0, 0), mask)
+        # Save file
+        white_image.save(diff_path)
+
+        # Count different pixels (black pixels in mask image are equal pixels)
+        equal_pixels = sum([1 for pixel in mask.getdata() if pixel == 0])
+        diff_pixels_percentage = 1 - equal_pixels / (baseline.width * baseline.height)
+        return diff_pixels_percentage
+
+    def _add_result_to_report(self, result, report_name, image_path, baseline_path, diff_path, message):
         """Add the result of a visual test to the html report
 
         :param result: comparation result (equal, diff, baseline)
         :param report_name: name to show in html report
-        :param image_file: image file path
-        :param baseline_file: baseline image file path
+        :param image_path: image file path
+        :param baseline_path: baseline image file path
+        :param diff_path: differences image file path
         :param message: error message
         """
         self.results[result] += 1
-        output_baseline_file = None
-        if baseline_file is not None:
-            output_baseline_file = os.path.join(self.output_directory, os.path.basename(baseline_file))
-            shutil.copyfile(baseline_file, output_baseline_file)
-        row = self._get_html_row(result, report_name, image_file, output_baseline_file, message)
+        output_baseline_path = None
+        if baseline_path is not None:
+            output_baseline_path = os.path.join(self.output_directory, os.path.basename(baseline_path))
+            shutil.copyfile(baseline_path, output_baseline_path)
+        row = self._get_html_row(result, report_name, image_path, output_baseline_path, diff_path, message)
         self._add_data_to_report_before_tag(row, '</tbody>')
         self._update_report_summary()
 
@@ -378,13 +406,14 @@ class VisualTest(object):
                                                                          self.results['diff'])
         self._add_data_to_report_before_tag(summary, '</div>')
 
-    def _get_html_row(self, result, report_name, image_file, baseline_file, message=''):
+    def _get_html_row(self, result, report_name, image_path, baseline_path, diff_path, message):
         """Create the html row with the result of a visual test
 
         :param result: comparation result (equal, diff, baseline)
         :param report_name: name to show in html report
-        :param image_file: image file path
-        :param baseline_file: baseline image file path
+        :param image_path: image file path
+        :param baseline_path: baseline image file path
+        :param diff_path: differences image file path
         :param message: error message
         :returns: str with the html row
         """
@@ -392,64 +421,33 @@ class VisualTest(object):
         row += '<td>' + report_name + '</td>'
 
         # Create baseline column
-        baseline_col = self._get_img_element(baseline_file, 'Baseline image')
+        baseline_col = self._get_img_element(baseline_path, 'Baseline image')
         row += '<td>' + baseline_col + '</td>'
 
         # Create image column
-        image_col = self._get_img_element(image_file, 'Screenshot image')
+        image_col = self._get_img_element(image_path, 'Screenshot image')
         row += '<td>' + image_col + '</td>'
 
         # Create diff column
-        diff_file = image_file.replace('.png', '.diff.png')
-        diff_col = self._get_img_element(diff_file, message) if os.path.exists(diff_file) else message
+        message = '' if message is None else message
+        diff_col = self._get_img_element(diff_path, message) if diff_path and os.path.exists(diff_path) else message
 
         row += '<td>' + diff_col + '</td>'
         row += '</tr>'
         return row
 
-    def _get_img_element(self, image_file, image_title):
+    def _get_img_element(self, image_path, image_title):
         """Create an img html element
 
-        :param image_file: filename of the image
+        :param image_path: image file path
         :param image_title: image title
         :returns: str with the img element
         """
-        if not image_file:
-            return ''
-        image_file_path = path.relpath(image_file, self.output_directory).replace('\\', '/')
-        return '<img src="{}" title="{}"/>'.format(image_file_path, image_title)
-
-    @staticmethod
-    def _get_diff_message(message, image_size):
-        """
-        Get formatted message with the distance between images
-
-        :param message: original engine message
-        :param image_size: number of pixels to convert absolute distances
-        :returns: formatted message
-        """
-        diff_message = None
-        if message is None:
-            # Images are equal
-            diff_message = ''
-        elif message == '' or 'Image dimensions do not match' in message:
-            # Different sizes in pil (''), perceptualdiff or imagemagick engines
-            diff_message = 'Image dimensions do not match'
-
-        if diff_message is None:
-            # Check pil or perceptualdiff engine message
-            m = re.search(r'\(by a distance of (.*)\)', message) or re.search(r'([0-9]*) pixels are different', message)
-            if m:
-                diff_message = 'Distance of %0.8f' % (float(m.group(1)) / image_size)
-        if diff_message is None:
-            # Check imagemagick engine message
-            m = re.search(r':[\r\n](\d*\.?\d*) \((\d*\.?\d*)\) @', message)
-            if m:
-                diff_message = 'Distance of %0.8f' % float(m.group(2))
-        if diff_message is None:
-            diff_message = message
-
-        return diff_message
+        img_element = ''
+        if image_path:
+            image_relative_path = path.relpath(image_path, self.output_directory).replace('\\', '/')
+            img_element = f'<img src="{image_relative_path}" title="{image_title}"/>'
+        return img_element
 
     @staticmethod
     def update_latest_report():